+++ CSD2b/eindopdracht/mix.cpp
@@ -3,20 +3,20 @@
 Mix::Mix(){}
 Mix::~Mix(){}
 
-float getA(float X){
+double getA(double X){
   return 1 - 0.25 * pow((X + 1), 2);
 }
-float getB(float X){
+double getB(double X){
   return 1 - pow((0.5X - 0.5), 2);
 }
 
-void setX(float X){
+void setX(double X){
   if(X >= 0. && X <= 0.99){
     this-> X = X;
   }else{
     this-> X = 0.1;
   }
 }
-float getX(){
+double getX(){
   return X;
 }
